requirements.txt
altair==4.0.0
argh==0.26.2
asgiref==3.2.3
astor==0.8.1
attrs==19.3.0
base58==1.0.3
blinker==1.4
boto3==1.10.46
botocore==1.13.46
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
Django==3.0.2
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
idna==2.8
image==1.5.27
importlib-metadata==1.3.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.0.2
numpy==1.18.0
pandas==0.25.3
pathtools==0.1.2
Pillow==7.0.0
protobuf==3.11.2
pyparsing==2.4.6
pyrsistent==0.15.6
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
streamlit==0.52.2
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tzlocal==2.0.0
urllib3==1.25.7
validators==0.14.1
watchdog==0.9.0
zipp==0.6.0
